Made Config class extensible and moved wire-frame loading into the OrientorConfig subclass. Moved the Config class into the car.shared.config package so that both UI's can use it. Added a fallback wire-frame for the OrientorConfig class.
Added code to the Config class which will attempt to read a form specification from the config file. Also added some code innt CarPicker and CarOrientor which will try to retrieve this information. The code will add a form and submit button which will submit the UI's important state information to the specified URL.

Added code into CarPicker and CarOrientor which will try to load the configuration file name from a global JavaScript variable, falling back to the default name if this fails. This gets around hard-wiring the configuration path and allows installation to run in different ways.

Moved some classes around to different packages for better organization.